Fix PingChurn
