+++ ipv8/test/peerdiscovery/test_churn.py
@@ -3,10 +3,8 @@ import time
 from ..base import TestBase
 from ..mocking.community import MockCommunity
 from ..mocking.endpoint import MockEndpointListener
-from ..mocking.ipv8 import MockIPv8
 from ...community import _DEFAULT_ADDRESSES
-from ...dht.community import DHTCommunity
-from ...peerdiscovery.churn import PingChurn, RandomChurn
+from ...peerdiscovery.churn import RandomChurn
 
 
 class TestChurn(TestBase):
@@ -123,46 +121,3 @@ class TestChurn(TestBase):
         await self.deliver_messages()
 
         self.assertEqual(len(sniffer.received_packets), 2)
-
-
-class TestPingChurn(TestBase):
-
-    def setUp(self):
-        super(TestPingChurn, self).setUp()
-
-        self.overlays = []
-        self.strategies = []
-
-        self.initialize(DHTCommunity, 2)
-
-    def create_node(self, *args, **kwargs):
-        peer = MockIPv8(u"low", DHTCommunity)
-
-        self.overlays.append(peer.overlay)
-        self.strategies.append(PingChurn(peer.overlay, ping_interval=0.0))
-
-        return peer
-
-    async def test_ping_all(self):
-        await self.introduce_nodes()
-        bucket = self.overlays[0].routing_table.trie[u'']
-
-        node1 = bucket.get(self.overlays[1].my_node_id)
-        node1.failed = 1
-        node1.last_response = 0
-
-        self.strategies[0].take_step()
-        await self.deliver_messages()
-
-        self.assertTrue(node1.failed == 0)
-        self.assertNotEqual(node1.last_response, 0)
-
-    async def test_ping_all_skip(self):
-        await self.introduce_nodes()
-        bucket = self.overlays[0].routing_table.trie[u'']
-        node1 = bucket.get(self.overlays[1].my_node_id)
-        node1.failed = 1
-        node1.last_response = time.time() + 5
-
-        self.strategies[0].take_step()
-        self.assertTrue(node1.failed == 1)
